% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.r
\name{inetrval_beta}
\alias{inetrval_beta}
\title{Power for Interval Hypothesis Tests}
\usage{
inetrval_beta(theta0, theta1, r, n, m, alpha, initial_guess)
}
\arguments{
\item{theta0}{$\theta_0$}

\item{theta1}{$\theta_1$}

\item{r}{$r$}

\item{n}{$n$}

\item{m}{$m$}

\item{alpha}{Significance level $\alpha$}
}
\value{
A probability value between 0 and 1
}
\description{
\code{interval_beta()} computes statistical power at $\theta = \theta_1$ for interval hypothesis tests.
}
