% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_alt.r
\name{interval_alt_critvals}
\alias{interval_alt_critvals}
\title{Critical Values for Interval-alternative UMP Test}
\usage{
interval_alt_critvals(theta1, theta2, r, n, m, alpha)
}
\arguments{
\item{theta1}{A positive numeric value representing θ₁}

\item{theta2}{A positive numeric value representing θ₂ (must be greater than θ₁)}

\item{r}{A non-zero numeric parameter of the scale-exponential family}

\item{n}{An integer representing the sample size}

\item{m}{A non-zero numeric parameter of the scale-exponential family}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}
}
\value{
A numeric length-two vector of critical values
}
\description{
\code{interval_alt_critvals} computes the critical values for the UMP test of
H0: θ ≤ θ₁ or θ ≥ θ₂ vs H1: θ₁ < θ < θ₂.
}
\examples{
interval_alt_critvals(1, sqrt(3), -2, 25, -1, 0.01)
}
