% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.r
\name{interval_critvals}
\alias{interval_critvals}
\title{Rejection Region for Interval Hypothesis Tests}
\usage{
interval_critvals(theta1, theta2, r, n, m, alpha)
}
\arguments{
\item{theta1}{A positive numeric value}

\item{theta2}{A positive numeric value}

\item{r}{A non-zero numeric value}

\item{n}{An integer represeting the sample size}

\item{m}{A non-zero numeric value}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}
}
\value{
A numeric vector
}
\description{
\code{interval_critvals()} computes the critival values for $H_0 : \theta \leq \theta_1 \text{ or } \theta \geq \theta_2$ vs $H_1 : \theta_1 < \theta < \theta_2$.
}
\examples{
interval_critvals(1, sqrt(3), -2, 25, -1, 0.01)
}
