% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_null.r
\name{interval_null_beta}
\alias{interval_null_beta}
\title{Power for Interval Hypothesis Tests}
\usage{
interval_null_beta(theta1, theta2, theta, r, n, m, alpha)
}
\arguments{
\item{theta1}{A positive numeric value}

\item{theta2}{A positive numeric value}

\item{theta}{A positive numeric value}

\item{r}{A non-zero numeric value}

\item{n}{An integer represeting the sample size}

\item{m}{A non-zero numeric value}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}
}
\value{
A numeric value between 0 and 1
}
\description{
\code{interval_null_beta()} computes statistical power at $\theta$ for interval hypothesis tests.
}
\examples{
interval_null_beta(1, sqrt(3), sqrt(2), -2, 25, -1, 0.01)
}
