% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.r
\name{interval_rr}
\alias{interval_rr}
\title{Rejection Region for Interval Hypothesis Tests}
\usage{
interval_rr(theta0, theta1, r, n, m, alpha, initial_guess)
}
\arguments{
\item{theta0}{$\theta_0$}

\item{r}{$r$}

\item{n}{$n$}

\item{m}{$m$}

\item{alpha}{Significance level $\alpha$}
}
\value{
A character vector
}
\description{
\code{interval_rr()} computes the rejection region for $H_0 : \theta \leq \theta_1 \text{ or } \theta \geq \theta_2$ vs $H_1 : \theta_1 < \theta < \theta_2$.
}
