% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.r
\name{simple_critval}
\alias{simple_critval}
\title{Critical Value for Simple MP and One-Sided UMP Tests}
\usage{
simple_critval(theta0, theta1, r, n, m, alpha)
}
\arguments{
\item{theta0}{A positive numeric value representing θ₀}

\item{theta1}{A positive numeric value representing θ₁}

\item{r}{A non-zero numeric parameter of the scale-exponential family}

\item{n}{An integer representing the sample size}

\item{m}{A non-zero numeric parameter of the scale-exponential family}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}
}
\value{
A numeric critical value
}
\description{
\code{simple_critval} computes the critical value for the MP test of
H₀: θ = θ₀ vs H₁: θ = θ₁, or UMP tests of
H₀: θ ≤ θ₀ vs H₁: θ > θ₀ and H₀: θ ≥ θ₀ vs H₁: θ < θ₀
}
\examples{
simple_critval(2, 4, -2, 20, -1, 0.01)
simple_critval(1 / 8, 1 / 10, -1, 9, -1, 0.05)
simple_critval(2, 1, 1, 1, 1, 0.05)
simple_critval(4, 6, 1, 1, 1, 0.05)
simple_critval(4, 2, 1, 1, 1, 0.05)
simple_critval(1.5, 3, -1, 10, -1, 0.05)
simple_critval(10, 14, -1, 31, -1, 0.05)
simple_critval(2, 1.25, -2, 40, -1, 0.025)
simple_critval(4, 12, -2, 25, -1, 0.05)
}
