% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.r
\name{simple_rr}
\alias{simple_rr}
\title{Rejection Region for Simple and One-Sided Hypotheses}
\usage{
simple_rr(theta0, theta1, r, n, m, alpha)
}
\arguments{
\item{theta0}{A positive numeric value}

\item{theta1}{A positive numeric value}

\item{r}{A non-zero numeric value}

\item{n}{An integer represeting the sample size}

\item{m}{A non-zero numeric value}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}
}
\value{
A character vector
}
\description{
\code{simple_rr()} computes the rejection region for simple vs simple,
simple vs one-sided and one-sided vs one-sided hypotheses.
}
\examples{
simple_rr(2, 4, -2, 20, -1, 0.01)
simple_rr(1 / 8, 1 / 10, -1, 9, -1, 0.05)
simple_rr(2, 1, 1, 1, 1, 0.05)
simple_rr(4, 6, 1, 1, 1, 0.05)
simple_rr(4, 2, 1, 1, 1, 0.05)
simple_rr(1.5, 3, -1, 10, -1, 0.05)
simple_rr(10, 14, -1, 31, -1, 0.05)
simple_rr(2, 1.25, -2, 40, -1, 0.025)
simple_rr(4, 12, -2, 25, -1, 0.05)
}
