% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.r
\name{two_sided_beta}
\alias{two_sided_beta}
\title{Power for Two-Sided Hypothesis Tests}
\usage{
two_sided_beta(theta0, theta1, r, n, m, alpha, initial_guess)
}
\arguments{
\item{theta0}{$\theta_0$}

\item{theta1}{$\theta_1$}

\item{r}{$r$}

\item{n}{$n$}

\item{m}{$m$}

\item{alpha}{Significance level $\alpha$}
}
\value{
A probability value between 0 and 1
}
\description{
\code{two_sided_beta()} computes statistical power at $\theta = \theta_1$ for simple vs two-sided hypotheses.
}
\examples{
two_sided_beta(1, -1, 1, -1, 0.05, c(0.05, 5))
two_sided_beta(1, -2, 4, -1, 0.05, c(0.3, 6))
}
