% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.r
\name{two_sided_critvals}
\alias{two_sided_critvals}
\title{Critical Value for Two-Sided UMPU Test}
\usage{
two_sided_critvals(theta0, r, n, m, alpha)
}
\arguments{
\item{theta0}{A positive numeric value representing θ₀}

\item{r}{A non-zero numeric parameter of the scale-exponential family}

\item{n}{An integer representing the sample size}

\item{m}{A non-zero numeric parameter of the scale-exponential family}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}
}
\value{
A numeric length-two vector of critical values
}
\description{
\code{two_sided_critvals} computes the critical values for the UMPU test of
H₀: θ = θ₀ vs H₁: θ ≠ θ₀.
}
\examples{
two_sided_critvals(1, -1, 1, -1, 0.05)
two_sided_critvals(1, -2, 4, -1, 0.05)
two_sided_critvals(5, -1, 500, -1, 0.05)
}
