% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.r
\name{two_sided_rr}
\alias{two_sided_rr}
\title{Rejection Region for Two-Sided Hypothesis Tests}
\usage{
two_sided_rr(theta0, r, n, m, alpha, initial_guess)
}
\arguments{
\item{theta0}{A positive numeric value}

\item{r}{A non-zero numeric value}

\item{n}{An integer represeting the sample size}

\item{m}{A non-zero numeric value}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}

\item{initial_guess}{Numeric vector of length 2}
}
\value{
A character vector
}
\description{
\code{two_sided_rr()} computes the rejection region for simple vs two-sided hypotheses.
}
\examples{
two_sided_rr(1, -1, 1, -1, 0.05, c(0.05, 5))
two_sided_rr(1, -2, 4, -1, 0.05, c(0.3, 6))
}
