% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.r
\name{inetrval_beta}
\alias{inetrval_beta}
\title{Power for Interval Hypothesis Tests}
\usage{
inetrval_beta(theta0, theta1, r, n, m, alpha, initial_guess)
}
\arguments{
\item{theta0}{A positive numeric value}

\item{theta1}{A positive numeric value}

\item{r}{A non-zero numeric value}

\item{n}{An integer represeting the sample size}

\item{m}{A non-zero numeric value}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}

\item{initial_guess}{Numeric vector of length 2}
}
\value{
A numeric value between 0 and 1
}
\description{
\code{interval_beta()} computes statistical power at $\theta = \theta_1$ for interval hypothesis tests.
}
