% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.r
\name{interval_rr}
\alias{interval_rr}
\title{Rejection Region for Interval Hypothesis Tests}
\usage{
interval_rr(theta0, theta1, r, n, m, alpha, initial_guess)
}
\arguments{
\item{theta0}{A positive numeric value}

\item{theta1}{A positive numeric value}

\item{r}{A non-zero numeric value}

\item{n}{An integer represeting the sample size}

\item{m}{A non-zero numeric value}

\item{alpha}{Numeric value between 0 and 1 representing the significance level}

\item{initial_guess}{Numeric vector of length 2}
}
\value{
A character vector
}
\description{
\code{interval_rr()} computes the rejection region for $H_0 : \theta \leq \theta_1 \text{ or } \theta \geq \theta_2$ vs $H_1 : \theta_1 < \theta < \theta_2$.
}
