% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.r
\name{simple_beta}
\alias{simple_beta}
\title{Power for Simple and One-Sided Hypotheses}
\usage{
simple_beta(theta0, theta1, r, n, m, alpha)
}
\arguments{
\item{theta0}{$\theta_0$}

\item{theta1}{$\theta_1$}

\item{r}{$r$}

\item{n}{$n$}

\item{m}{$m$}

\item{alpha}{Significance level $\alpha$}
}
\value{
A probability value between 0 and 1
}
\description{
\code{simple_beta()} computes statistical power at $\theta = \theta_1$ for simple vs simple,
simple vs one-sided and one-sided vs one-sided tests.
}
\examples{
simple_beta(3, 5, -2, 25, -1, 0.05)
simple_beta(3, 2, -2, 25, -1, 0.05)
simple_beta(4, 12, -2, 25, -1, 0.05)
simple_beta(2, 12, -2, 28, -1, 0.05)
}
