% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.r
\name{simple_rr}
\alias{simple_rr}
\title{Rejection Region for Simple and One-Sided Hypotheses}
\usage{
simple_rr(theta0, theta1, r, n, m, alpha)
}
\arguments{
\item{theta0}{$\theta_0$}

\item{theta1}{$\theta_1$}

\item{r}{$r$}

\item{n}{$n$}

\item{m}{$m$}

\item{alpha}{Significance level $\alpha$}
}
\value{
A character vector
}
\description{
\code{simple_rr()} computes the rejection region for simple vs simple,
simple vs one-sided and one-sided vs one-sided hypotheses.
}
\examples{
simple_rr(3, 5, -2, 25, -1, 0.05)
simple_rr(3, 2, -2, 25, -1, 0.05)
simple_rr(4, 12, -2, 25, -1, 0.05)
simple_rr(2, 12, -2, 28, -1, 0.05)
}
