% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.r
\name{two_sided_rr}
\alias{two_sided_rr}
\title{Rejection Region for Two-Sided Hypothesis Tests}
\usage{
two_sided_rr(theta0, r, n, m, alpha, initial_guess)
}
\arguments{
\item{theta0}{$\theta_0$}

\item{r}{$r$}

\item{n}{$n$}

\item{m}{$m$}

\item{alpha}{Significance level $\alpha$}
}
\value{
A character vector
}
\description{
\code{two_sided_rr()} computes the rejection region for simple vs two-sided hypotheses.
}
\examples{
two_sided_rr(1, -1, 1, -1, 0.05, c(0.05, 5))
two_sided_rr(1, -2, 4, -1, 0.05, c(0.3, 6))
}
